+++ MiniApp/Classes/JavascriptBridge/MiniAppScriptMessageHandler.swift
@@ -16,30 +16,32 @@ internal class MiniAppScriptMessageHandler: NSObject, WKScriptMessageHandler {
     weak var delegate: MiniAppCallbackDelegate?
     weak var hostAppMessageDelegate: MiniAppMessageDelegate?
     var miniAppId: String
+    var miniAppTitle: String
     var userAlreadyRespondedRequestList = [MASDKCustomPermissionModel]()
     var cachedUnknownCustomPermissionRequest = [MiniAppCustomPermissionsListResponse]()
     var miniAppKeyStore = MiniAppKeyChain()
 
-    init(delegate: MiniAppCallbackDelegate, hostAppMessageDelegate: MiniAppMessageDelegate, miniAppId: String) {
+    init(delegate: MiniAppCallbackDelegate, hostAppMessageDelegate: MiniAppMessageDelegate, miniAppId: String, miniAppTitle: String) {
         self.delegate = delegate
         self.hostAppMessageDelegate = hostAppMessageDelegate
         self.miniAppId = miniAppId
+        self.miniAppTitle = miniAppTitle
         super.init()
     }
 
     func userContentController(_ userContentController: WKUserContentController, didReceive message: WKScriptMessage) {
-            if let messageBody = message.body as? String {
-                let bodyData: Data = messageBody.data(using: .utf8)!
-                let responseJson = ResponseDecoder.decode(decodeType: MiniAppJavaScriptMessageInfo.self, data: bodyData)
-                handleBridgeMessage(responseJson: responseJson)
-            }
+        if let messageBody = message.body as? String {
+            let bodyData: Data = messageBody.data(using: .utf8)!
+            let responseJson = ResponseDecoder.decode(decodeType: MiniAppJavaScriptMessageInfo.self, data: bodyData)
+            handleBridgeMessage(responseJson: responseJson)
+        }
     }
 
     func handleBridgeMessage(responseJson: MiniAppJavaScriptMessageInfo?) {
         guard let actionCommand = responseJson?.action, !actionCommand.isEmpty,
             let callbackId = responseJson?.id, !callbackId.isEmpty, let requestAction = MiniAppJSActionCommand(rawValue: actionCommand) else {
                 executeJavaScriptCallback(responseStatus: .onError, messageId: "", response: getMiniAppErrorMessage(MiniAppJavaScriptError.unexpectedMessageFormat))
-            return
+                return
         }
         let requestParam = responseJson?.param ?? nil
         handleActionCommand(action: requestAction, requestParam: requestParam, callbackId: callbackId)
@@ -62,6 +64,10 @@ internal class MiniAppScriptMessageHandler: NSObject, WKScriptMessageHandler {
             fetchUserName(callbackId: callbackId)
         case .getProfilePhoto:
             fetchProfilePhoto(callbackId: callbackId)
+        case .setScreenOrientation:
+            setScreenOrientation(requestParam: requestParam, callbackId: callbackId)
+        case .getAccessToken:
+            fetchTokenDetails(callbackId: callbackId)
         }
     }
 
@@ -105,14 +111,14 @@ internal class MiniAppScriptMessageHandler: NSObject, WKScriptMessageHandler {
     }
 
     func getLocationInfo() -> String {
-        return  "{\"coords\":{\"latitude\":\(locationManager?.location?.coordinate.latitude ?? 0)" +
-        ",\"longitude\":\(locationManager?.location?.coordinate.longitude ?? 0)" +
-        ", \"altitude\":\(locationManager?.location?.altitude ?? "null" as Any)" +
-        ", \"altitudeAccuracy\":\(locationManager?.location?.verticalAccuracy ?? "null" as Any)" +
-        ", \"accuracy\":\(locationManager?.location?.horizontalAccuracy ?? 0)" +
-        ", \"speed\":\(locationManager?.location?.speed ?? "null" as Any)" +
-        ", \"heading\":\(locationManager?.location?.course ?? "null" as Any)" +
-        "}, \"timestamp\":\(Date().epochInMilliseconds)}"
+        return "{\"coords\":{\"latitude\":\(locationManager?.location?.coordinate.latitude ?? 0)" +
+            ",\"longitude\":\(locationManager?.location?.coordinate.longitude ?? 0)" +
+            ", \"altitude\":\(locationManager?.location?.altitude ?? "null" as Any)" +
+            ", \"altitudeAccuracy\":\(locationManager?.location?.verticalAccuracy ?? "null" as Any)" +
+            ", \"accuracy\":\(locationManager?.location?.horizontalAccuracy ?? 0)" +
+            ", \"speed\":\(locationManager?.location?.speed ?? "null" as Any)" +
+            ", \"heading\":\(locationManager?.location?.course ?? "null" as Any)" +
+            "}, \"timestamp\":\(Date().epochInMilliseconds)}"
     }
 
     func executeJavaScriptCallback(responseStatus: JavaScriptExecResult, messageId: String, response: String) {
@@ -180,6 +186,39 @@ internal class MiniAppScriptMessageHandler: NSObject, WKScriptMessageHandler {
     func isUserAllowedPermission(customPermissionType: MiniAppCustomPermissionType) -> Bool {
         let customPermission = self.miniAppKeyStore.getCustomPermissions(forMiniApp: self.miniAppId).filter { $0.permissionName == customPermissionType }
         return customPermission[0].isPermissionGranted.boolValue
+    }
+
+    func setScreenOrientation(requestParam: RequestParameters?, callbackId: String) {
+        guard let requestParamValue = requestParam?.action else {
+            executeJavaScriptCallback(responseStatus: .onError, messageId: callbackId, response: getMiniAppErrorMessage(MiniAppJavaScriptError.unexpectedMessageFormat))
+            return
+        }
+        if !requestParamValue.isEmpty {
+            guard let info = MiniAppInterfaceOrientation(rawValue: requestParamValue) else {
+                self.executeJavaScriptCallback(responseStatus: .onError, messageId: callbackId, response: getMiniAppErrorMessage(MiniAppJavaScriptError.unexpectedMessageFormat))
+                return
+            }
+            MiniApp.MAOrientationLock = info.orientation
+            UIViewController.attemptRotationToDeviceOrientation()
+            executeJavaScriptCallback(responseStatus: .onSuccess, messageId: callbackId, response: MASDKProtocolResponse.success.rawValue)
+        } else {
+            self.executeJavaScriptCallback(responseStatus: .onError, messageId: callbackId, response: getMiniAppErrorMessage(MiniAppJavaScriptError.valueIsEmpty))
+        }
+    }
+
+    func fetchTokenDetails(callbackId: String) {
+        hostAppMessageDelegate?.getAccessToken(miniAppId: self.miniAppId) { (result) in
+            switch result {
+            case .success(let responseMessage):
+                guard let jsonResponse = ResponseEncoder.encode(data: responseMessage) else {
+                    self.executeJavaScriptCallback(responseStatus: .onError, messageId: callbackId, response: getMiniAppErrorMessage(MiniAppJavaScriptError.internalError))
+                    return
+                }
+                self.executeJavaScriptCallback(responseStatus: .onSuccess, messageId: callbackId, response: jsonResponse)
+            case .failure(let error):
+                self.executeJavaScriptCallback(responseStatus: .onError, messageId: callbackId, response: getMiniAppErrorMessage(error))
+            }
+        }
     }
 }