+++ MiniApp/Classes/Protocols/MiniAppDisplayProtocol.swift
@@ -3,8 +3,11 @@
  to communicate with the Mini App display module
  */
 
-public protocol MiniAppDisplayProtocol {
+public typealias MiniAppDisplayProtocol = MiniAppDisplayDelegate
+
+public protocol MiniAppDisplayDelegate: class {
 
     /// Get the view of the Mini app
     func getMiniAppView() -> UIView
+
 }