+++ Tests/Unit/MiniAppScriptMessageHandlerTests.swift
@@ -4,6 +4,8 @@ import WebKit
 @testable import MiniApp
 
 // swiftlint:disable function_body_length
+// swiftlint:disable file_length
+// swiftlint:disable type_body_length
 class MiniAppScriptMessageHandlerTests: QuickSpec {
 
     override func spec() {
@@ -14,7 +16,7 @@ class MiniAppScriptMessageHandlerTests: QuickSpec {
             let scriptMessageHandler = MiniAppScriptMessageHandler(
                 delegate: callbackProtocol,
                 hostAppMessageDelegate: mockMessageInterface,
-                miniAppId: "Test"
+                miniAppId: "Test", miniAppTitle: "Mini App"
             )
             afterEach {
                 deleteStatusPreferences()
@@ -34,7 +36,7 @@ class MiniAppScriptMessageHandlerTests: QuickSpec {
                     let scriptMessageHandler = MiniAppScriptMessageHandler(
                         delegate: callbackProtocol,
                         hostAppMessageDelegate: mockMessageInterface,
-                        miniAppId: "Test"
+                        miniAppId: "Test", miniAppTitle: "Mini App"
                     )
                     let mockMessage = MockWKScriptMessage(name: "", body: "{\"action\": \"getUniqueId\", \"param\": { \"permission\": null}, \"id\":\"12345\"}" as AnyObject)
                     scriptMessageHandler.userContentController(WKUserContentController(), didReceive: mockMessage)
@@ -48,9 +50,9 @@ class MiniAppScriptMessageHandlerTests: QuickSpec {
                     let scriptMessageHandler = MiniAppScriptMessageHandler(
                         delegate: callbackProtocol,
                         hostAppMessageDelegate: mockMessageInterface,
-                        miniAppId: "Test"
+                        miniAppId: "Test", miniAppTitle: "Mini App"
                     )
-                    let requestParam = RequestParameters(permission: "location", permissions: nil, locationOptions: nil, shareInfo: ShareInfoParameters(content: ""))
+                    let requestParam = RequestParameters(action: "", permission: "location", permissions: nil, locationOptions: nil, shareInfo: ShareInfoParameters(content: ""))
                     let javascriptMessageInfo = MiniAppJavaScriptMessageInfo(action: "", id: "123", param: requestParam)
                     scriptMessageHandler.handleBridgeMessage(responseJson: javascriptMessageInfo)
                     expect(callbackProtocol.errorMessage).toEventually(contain(MiniAppJavaScriptError.unexpectedMessageFormat.rawValue))
@@ -61,9 +63,9 @@ class MiniAppScriptMessageHandlerTests: QuickSpec {
                     let scriptMessageHandler = MiniAppScriptMessageHandler(
                         delegate: callbackProtocol,
                         hostAppMessageDelegate: mockMessageInterface,
-                        miniAppId: "Test"
+                        miniAppId: "Test", miniAppTitle: "Mini App"
                     )
-                    let requestParam = RequestParameters(permission: "location", permissions: nil, locationOptions: nil, shareInfo: ShareInfoParameters(content: ""))
+                    let requestParam = RequestParameters(action: "", permission: "location", permissions: nil, locationOptions: nil, shareInfo: ShareInfoParameters(content: ""))
                     let javascriptMessageInfo = MiniAppJavaScriptMessageInfo(action: "getUniqueId", id: "", param: requestParam)
                     scriptMessageHandler.handleBridgeMessage(responseJson: javascriptMessageInfo)
                     expect(callbackProtocol.errorMessage).toEventually(contain(MiniAppJavaScriptError.unexpectedMessageFormat.rawValue))
@@ -75,7 +77,7 @@ class MiniAppScriptMessageHandlerTests: QuickSpec {
                     let scriptMessageHandler = MiniAppScriptMessageHandler(
                         delegate: callbackProtocol,
                         hostAppMessageDelegate: mockMessageInterface,
-                        miniAppId: "Test"
+                        miniAppId: "Test", miniAppTitle: "Mini App"
                     )
                     scriptMessageHandler.sendUniqueId(messageId: "1234")
                     expect(callbackProtocol.errorMessage).toEventually(contain(MiniAppJavaScriptError.internalError.rawValue))
@@ -87,7 +89,7 @@ class MiniAppScriptMessageHandlerTests: QuickSpec {
                      let scriptMessageHandler = MiniAppScriptMessageHandler(
                         delegate: callbackProtocol,
                         hostAppMessageDelegate: mockMessageInterface,
-                        miniAppId: "Test"
+                        miniAppId: "Test", miniAppTitle: "Mini App"
                     )
                     mockMessageInterface.locationAllowed = true
                     let mockMessage = MockWKScriptMessage(name: "", body: "{\"action\": \"requestPermission\", \"param\": { \"permission\": \"location\"}, \"id\":\"12345\"}" as AnyObject)
@@ -100,7 +102,7 @@ class MiniAppScriptMessageHandlerTests: QuickSpec {
                      let scriptMessageHandler = MiniAppScriptMessageHandler(
                         delegate: callbackProtocol,
                         hostAppMessageDelegate: mockMessageInterface,
-                        miniAppId: "Test"
+                        miniAppId: "Test", miniAppTitle: "Mini App"
                     )
                     mockMessageInterface.locationAllowed = false
                     let mockMessage = MockWKScriptMessage(name: "", body: "{\"action\": \"requestPermission\", \"param\": { \"permission\": \"location\"}, \"id\":\"12345\"}" as AnyObject)
@@ -113,7 +115,7 @@ class MiniAppScriptMessageHandlerTests: QuickSpec {
                      let scriptMessageHandler = MiniAppScriptMessageHandler(
                         delegate: callbackProtocol,
                         hostAppMessageDelegate: mockMessageInterface,
-                        miniAppId: "Test"
+                        miniAppId: "Test", miniAppTitle: "Mini App"
                     )
                     mockMessageInterface.locationAllowed = false
                     mockMessageInterface.permissionError = .notDetermined
@@ -127,7 +129,7 @@ class MiniAppScriptMessageHandlerTests: QuickSpec {
                      let scriptMessageHandler = MiniAppScriptMessageHandler(
                         delegate: callbackProtocol,
                         hostAppMessageDelegate: mockMessageInterface,
-                        miniAppId: "Test"
+                        miniAppId: "Test", miniAppTitle: "Mini App"
                     )
                     mockMessageInterface.locationAllowed = false
                     mockMessageInterface.permissionError = .notDetermined
@@ -141,7 +143,7 @@ class MiniAppScriptMessageHandlerTests: QuickSpec {
                      let scriptMessageHandler = MiniAppScriptMessageHandler(
                         delegate: callbackProtocol,
                         hostAppMessageDelegate: mockMessageInterface,
-                        miniAppId: "Test"
+                        miniAppId: "Test", miniAppTitle: "Mini App"
                     )
                     mockMessageInterface.locationAllowed = false
                     let mockMessage = MockWKScriptMessage(name: "", body: "{\"action\": \"getCurrentPosition\", \"param\":null, \"id\":\"12345\"}" as AnyObject)
@@ -154,7 +156,7 @@ class MiniAppScriptMessageHandlerTests: QuickSpec {
                      let scriptMessageHandler = MiniAppScriptMessageHandler(
                         delegate: callbackProtocol,
                         hostAppMessageDelegate: mockMessageInterface,
-                        miniAppId: "Test"
+                        miniAppId: "Test", miniAppTitle: "Mini App"
                     )
                     mockMessageInterface.locationAllowed = false
                     let mockMessage = MockWKScriptMessage(name: "", body: "{\"action\": \"getCurrentPosition\", \"param\":null, \"id\":\"12345\"}" as AnyObject)
@@ -167,7 +169,7 @@ class MiniAppScriptMessageHandlerTests: QuickSpec {
                      let scriptMessageHandler = MiniAppScriptMessageHandler(
                         delegate: callbackProtocol,
                         hostAppMessageDelegate: mockMessageInterface,
-                        miniAppId: "Test"
+                        miniAppId: "Test", miniAppTitle: "Mini App"
                     )
                     mockMessageInterface.customPermissions = true
                     let mockMessage = MockWKScriptMessage(
@@ -188,7 +190,7 @@ class MiniAppScriptMessageHandlerTests: QuickSpec {
                      let scriptMessageHandler = MiniAppScriptMessageHandler(
                         delegate: callbackProtocol,
                         hostAppMessageDelegate: mockMessageInterface,
-                        miniAppId: "Test"
+                        miniAppId: "Test", miniAppTitle: "Mini App"
                     )
                     mockMessageInterface.customPermissions = true
                     let mockMessage = MockWKScriptMessage(
@@ -204,7 +206,7 @@ class MiniAppScriptMessageHandlerTests: QuickSpec {
                      let scriptMessageHandler = MiniAppScriptMessageHandler(
                         delegate: callbackProtocol,
                         hostAppMessageDelegate: mockMessageInterface,
-                        miniAppId: "Test"
+                        miniAppId: "Test", miniAppTitle: "Mini App"
                     )
                     mockMessageInterface.customPermissions = true
                     let mockMessage = MockWKScriptMessage(
@@ -220,7 +222,7 @@ class MiniAppScriptMessageHandlerTests: QuickSpec {
                      let scriptMessageHandler = MiniAppScriptMessageHandler(
                         delegate: callbackProtocol,
                         hostAppMessageDelegate: mockMessageInterface,
-                        miniAppId: "Test"
+                        miniAppId: "Test", miniAppTitle: "Mini App"
                     )
                     mockMessageInterface.customPermissions = true
                     let mockMessage = MockWKScriptMessage(
@@ -233,7 +235,7 @@ class MiniAppScriptMessageHandlerTests: QuickSpec {
             context("when MiniAppScriptMessageHandler receives valid share info message") {
                 it("will return SUCCESS after host app displays UIActivityController") {
                     mockMessageInterface.messageContentAllowed = true
-                    let scriptMessageHandler = MiniAppScriptMessageHandler(delegate: callbackProtocol, hostAppMessageDelegate: mockMessageInterface, miniAppId: "Test")
+                    let scriptMessageHandler = MiniAppScriptMessageHandler(delegate: callbackProtocol, hostAppMessageDelegate: mockMessageInterface, miniAppId: "Test", miniAppTitle: "Mini App")
                     let mockMessage = MockWKScriptMessage(
                         name: "", body: "{\"action\":\"shareInfo\",\"param\":{\"shareInfo\":{\"content\":\"Test Message\"}},\"id\":\"1.033890137027198\"}"as AnyObject)
                     scriptMessageHandler.userContentController(WKUserContentController(), didReceive: mockMessage)
@@ -244,7 +246,7 @@ class MiniAppScriptMessageHandlerTests: QuickSpec {
             context("when MiniAppScriptMessageHandler receives valid share info message but there is an error in host app delegate") {
                 it("will return Error") {
                     mockMessageInterface.messageContentAllowed = false
-                    let scriptMessageHandler = MiniAppScriptMessageHandler(delegate: callbackProtocol, hostAppMessageDelegate: mockMessageInterface, miniAppId: "Test")
+                    let scriptMessageHandler = MiniAppScriptMessageHandler(delegate: callbackProtocol, hostAppMessageDelegate: mockMessageInterface, miniAppId: "Test", miniAppTitle: "Mini App")
                     let mockMessage = MockWKScriptMessage(
                         name: "", body: "{\"action\":\"shareInfo\",\"param\":{\"shareInfo\":{\"content\":\"Test Message\"}},\"id\":\"1.033890137027198\"}"as AnyObject)
                     scriptMessageHandler.userContentController(WKUserContentController(), didReceive: mockMessage)
@@ -259,7 +261,7 @@ class MiniAppScriptMessageHandlerTests: QuickSpec {
                      let scriptMessageHandler = MiniAppScriptMessageHandler(
                         delegate: mockCallbackProtocol,
                         hostAppMessageDelegate: mockMessageInterface,
-                        miniAppId: "Test"
+                        miniAppId: "Test", miniAppTitle: "Mini App"
                     )
                     mockMessageInterface.messageContentAllowed = false
                     let mockMessage = MockWKScriptMessage(
@@ -274,7 +276,7 @@ class MiniAppScriptMessageHandlerTests: QuickSpec {
                     let scriptMessageHandler = MiniAppScriptMessageHandler(
                        delegate: mockCallbackProtocol,
                        hostAppMessageDelegate: mockMessageInterface,
-                       miniAppId: self.name
+                       miniAppId: self.name, miniAppTitle: "Mini App"
                     )
                     updateCustomPermissionStatus(miniAppId: self.name, permissionType: .userName, status: .allowed)
                    let mockMessage = MockWKScriptMessage(
@@ -289,7 +291,7 @@ class MiniAppScriptMessageHandlerTests: QuickSpec {
                     let scriptMessageHandler = MiniAppScriptMessageHandler(
                        delegate: mockCallbackProtocol,
                        hostAppMessageDelegate: mockMessageInterface,
-                       miniAppId: self.name
+                       miniAppId: self.name, miniAppTitle: "Mini App"
                     )
                     updateCustomPermissionStatus(miniAppId: self.name, permissionType: .userName, status: .denied)
                    let mockMessage = MockWKScriptMessage(
@@ -305,7 +307,7 @@ class MiniAppScriptMessageHandlerTests: QuickSpec {
                      let scriptMessageHandler = MiniAppScriptMessageHandler(
                         delegate: mockCallbackProtocol,
                         hostAppMessageDelegate: mockMessageInterface,
-                        miniAppId: self.name
+                        miniAppId: self.name, miniAppTitle: "Mini App"
                     )
                     updateCustomPermissionStatus(miniAppId: self.name, permissionType: .profilePhoto, status: .allowed)
                     let mockMessage = MockWKScriptMessage(
@@ -318,7 +320,7 @@ class MiniAppScriptMessageHandlerTests: QuickSpec {
                     let scriptMessageHandler = MiniAppScriptMessageHandler(
                        delegate: mockCallbackProtocol,
                        hostAppMessageDelegate: mockMessageInterface,
-                       miniAppId: self.name
+                       miniAppId: self.name, miniAppTitle: "Mini App"
                    )
                    updateCustomPermissionStatus(miniAppId: self.name, permissionType: .profilePhoto, status: .allowed)
                     mockMessageInterface.mockProfilePhoto = ""
@@ -332,7 +334,7 @@ class MiniAppScriptMessageHandlerTests: QuickSpec {
                     let scriptMessageHandler = MiniAppScriptMessageHandler(
                        delegate: mockCallbackProtocol,
                        hostAppMessageDelegate: mockMessageInterface,
-                       miniAppId: self.name
+                       miniAppId: self.name, miniAppTitle: "Mini App"
                    )
                    updateCustomPermissionStatus(miniAppId: self.name, permissionType: .profilePhoto, status: .denied)
                    let mockMessage = MockWKScriptMessage(
@@ -340,6 +342,66 @@ class MiniAppScriptMessageHandlerTests: QuickSpec {
                    scriptMessageHandler.userContentController(WKUserContentController(), didReceive: mockMessage)
                    expect(mockCallbackProtocol.errorMessage).toEventually(contain(MASDKCustomPermissionError.userDenied.rawValue), timeout: .seconds(10))
                 }
+            }
+            context("when MiniAppScriptMessageHandler receives valid getAccessToken command") {
+                 it("will return Access Token data info such as Token string and expiry date") {
+                    let mockCallbackProtocol = MockMiniAppCallbackProtocol()
+                     let scriptMessageHandler = MiniAppScriptMessageHandler(
+                        delegate: mockCallbackProtocol,
+                        hostAppMessageDelegate: mockMessageInterface,
+                        miniAppId: self.name, miniAppTitle: "Mini App"
+                    )
+                    mockMessageInterface.mockAccessToken = true
+                    let mockMessage = MockWKScriptMessage(
+                        name: "", body: "{\"action\": \"getAccessToken\", \"param\":null, \"id\":\"12345\"}" as AnyObject)
+                    scriptMessageHandler.userContentController(WKUserContentController(), didReceive: mockMessage)
+                    expect(mockCallbackProtocol.response).toEventually(contain("MOCK_ACCESS_TOKEN"), timeout: .seconds(10))
+                 }
+            }
+            context("when MiniAppScriptMessageHandler receives valid getAccessToken command but host app returns error") {
+                 it("will return error") {
+                    let mockCallbackProtocol = MockMiniAppCallbackProtocol()
+                     let scriptMessageHandler = MiniAppScriptMessageHandler(
+                        delegate: mockCallbackProtocol,
+                        hostAppMessageDelegate: mockMessageInterface,
+                        miniAppId: self.name, miniAppTitle: "Mini App"
+                    )
+                    mockMessageInterface.mockAccessToken = false
+                    let mockMessage = MockWKScriptMessage(
+                        name: "", body: "{\"action\": \"getAccessToken\", \"param\":null, \"id\":\"12345\"}" as AnyObject)
+                    scriptMessageHandler.userContentController(WKUserContentController(), didReceive: mockMessage)
+                    expect(mockCallbackProtocol.errorMessage).toEventually(contain(MASDKCustomPermissionError.unknownError.rawValue), timeout: .seconds(10))
+                 }
+            }
+            context("when MiniAppScriptMessageHandler receives invalid orientation lock command") {
+                 it("will return error") {
+                    let mockCallbackProtocol = MockMiniAppCallbackProtocol()
+                    let scriptMessageHandler = MiniAppScriptMessageHandler(
+                        delegate: mockCallbackProtocol,
+                        hostAppMessageDelegate: mockMessageInterface,
+                        miniAppId: "Test", miniAppTitle: "Mini App"
+                    )
+                    mockMessageInterface.messageContentAllowed = false
+                    let mockMessage = MockWKScriptMessage(
+                        name: "", body: "{\"action\":\"setScreenOrientation\",\"param\":{\"action\":null},\"id\":\"5.733550049709592\"}" as AnyObject)
+                    scriptMessageHandler.userContentController(WKUserContentController(), didReceive: mockMessage)
+                    expect(mockCallbackProtocol.errorMessage).toEventually(contain(MiniAppJavaScriptError.unexpectedMessageFormat.rawValue), timeout: .seconds(10))
+                 }
+            }
+            context("when MiniAppScriptMessageHandler receives no orientation lock command") {
+                 it("will return error") {
+                    let mockCallbackProtocol = MockMiniAppCallbackProtocol()
+                    let scriptMessageHandler = MiniAppScriptMessageHandler(
+                        delegate: mockCallbackProtocol,
+                        hostAppMessageDelegate: mockMessageInterface,
+                        miniAppId: "Test", miniAppTitle: "Mini App"
+                    )
+                    mockMessageInterface.messageContentAllowed = false
+                    let mockMessage = MockWKScriptMessage(
+                        name: "", body: "{\"action\":\"setScreenOrientation\",\"param\":{\"action\":\"rakuten.miniapp.screen.LOCK_LANDSCAPE_RIGHT\"},\"id\":\"5.733550049709592\"}" as AnyObject)
+                    scriptMessageHandler.userContentController(WKUserContentController(), didReceive: mockMessage)
+                    expect(mockCallbackProtocol.errorMessage).toEventually(contain(MiniAppJavaScriptError.unexpectedMessageFormat.rawValue), timeout: .seconds(10))
+                 }
             }
         }
     }